Backend/requirements.txt
torch~=2.4.1
Django~=5.1.4
djangorestframework~=3.15.2